splinting with new rules now gets up to hdf5 module

+++ scilab/modules/sparse/includes/spmalloc.h
@@ -1,11 +1,11 @@
 /*
  * Scilab ( http://www.scilab.org/ ) - This file is part of Scilab
  * Copyright (C) INRIA
- * 
+ *
  * This file must be used under the terms of the CeCILL.
  * This source file is licensed as described in the file COPYING, which
  * you should have received as part of this distribution.  The terms
- * are also available at    
+ * are also available at
  * http://www.cecill.info/licences/Licence_CeCILL_V2-en.txt
  *
  */
@@ -16,9 +16,9 @@
 #include <stdlib.h>
 #include "MALLOC.h"
 #define SPMALLOC(x) MALLOC(((size_t) x))
-#define SPALLOC(type,number)  ((type *)SPMALLOC((unsigned)(sizeof(type)*(number))))
+#define SPALLOC(type,number)  ((type *)SPMALLOC((sizeof(type)*(number))))
 #define SPREALLOC(ptr,type,number)  \
-           ptr = (type *)REALLOC((char *)ptr,(unsigned)(sizeof(type)*(number)))
+           ptr = (type *)REALLOC((char *)ptr,(sizeof(type)*(number)))
 
 #define SPFREE(ptr) { if ((ptr) != NULL) {FREE((void *)(ptr)); (ptr) = NULL;}}
